Hooked up flowrate data to a firebase realtime database. Changed textrue/fixed caustics on bucket simulation, changed texture to plastic vs. stone (for now, at least). Hooked up flowrate to bucket simulation, as well. Setting a global maxiumum at 20,000 cubic feet/second, which is then overwritten if the flowpath has a higher max value.
