+++ .gitignore
@@ -6,6 +6,7 @@
 # Data files too large for github storage
 /public/data/rivers_full_detail.json
 /public/data/detailed_rivers_streams.json
+/public/data/flowrates.json
 
 # Data files too large for github storage
 # Original VAA .parquet file can be found here: https://www.hydroshare.org/resource/6092c8a62fac45be97a09bfd0b0bf726/
